pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "src"
version = "0.0.1"
description = "Following along with Sebastian Raschka's book."
authors = [{name = "Hamed Bastan-Hagh", email = "hamed@bastanhagh.com" }]
readme = "README.md"
requires-python = ">= 3.10"
dependencies = [
    "click",
    "ipykernel",
    "nbstripout",
    "numpy",
    "python-dotenv",
    "pickleshare",
    "tiktoken",
    "torch",
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "pytest",
    "ruff",
]

[tool.ruff]
line-length = 79
indent-width = 4

[tool.ruff.lint]
# The default set of rules is "E4", "E7", "E9", "F"
select = ["E4", "E7", "E9", "F", "I"]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = true

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"
